dmaDataTx_prepare() now can fail because DMA is still busy. Checking of WDSR returned back, that was just bad WORD pointer where BYTE pointer should be.
git-svn-id: https://subversion.assembla.com/svn/atarijookie/trunk@953 0d2ca917-fb58-4453-bb26-a698c01b0940

+++ test_manual/scsi.h
@@ -88,7 +88,7 @@
 
 #define WDC                 ((volatile WORD *)  0xFFFF8604)
 #define WDL                 ((volatile WORD *)  0xFFFF8606)
-#define WDSR                ((volatile WORD *)  0xFFFF860F)       // STATUS Register
+#define WDSR                ((volatile BYTE *)  0xFFFF860F)       // STATUS Register
 
 // Falcon regs for READing
 #define SPCSD           0x88    // SPCSD      - R  Current SCSI Data
